% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_alpha_names.R
\name{cr_alpha_names}
\alias{cr_alpha_names}
\title{Provide names for Cost Report "Alpha Table" data}
\usage{
cr_alpha_names()
}
\value{
A list of names for the cost report Alpha Table
}
\description{
Provide names for Cost Report "Alpha Table" data
}
\examples{

# get the list
cr_nmrc_names()

}

