% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_nmrc_names.R
\name{cr_nmrc_names}
\alias{cr_nmrc_names}
\title{Provide names for Cost Report "Numeric Table" data}
\usage{
cr_nmrc_names()
}
\value{
A list of names for the cost report Numeric Table
}
\description{
Provide names for Cost Report "Numeric Table" data
}
\examples{

# get the list
cr_nmrc_names()

}
