% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_rpt_names.R
\name{cr_rpt_names}
\alias{cr_rpt_names}
\title{Provide names for Cost Report "Report Table" data}
\usage{
cr_rpt_names()
}
\value{
A list of names for the cost report Report Table
}
\description{
Provide names for Cost Report "Report Table" data
}
\examples{

# get the list
cr_rpt_names()

}

