% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare.R
\docType{package}
\name{medicare}
\alias{medicare}
\alias{medicare-package}
\title{medicare: tools for simplifying Medicare data analysis}
\description{
The medicare package contains useful functions for manipulating
  raw Medicare public use files. These sometimes come with SAS read-in code
  provided, but more frequently require the analyst to manually recode and
  rename variables based on thorough review of the data documentation. This
  package focuses mostly on Cost Reports and Provider of Service files, but
  more support will be added for other sources in the future.
}

