% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pos2010}
\alias{pos2010}
\title{Sample Medicare Provider of Service 2010 data for hospices}
\format{
A data frame with 402 rows and 530 variables
}
\source{
\url{https://www.nber.org/data/provider-of-services.html}
}
\usage{
pos2010
}
\description{
A sample of Provider of Services data for select hospices in 2010. 
This is raw data, similar to what you'd get on your own when reading in a csv.
}
\keyword{datasets}
