% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_column.R
\name{subset_column}
\alias{subset_column}
\title{Subset the desired dataset, based on a column code.}
\usage{
subset_column(dataset, column)
}
\arguments{
\item{dataset}{The name of a cost report alpha or numeric dataset}

\item{column}{The column of the workheet}
}
\value{
A subset of the provided dataset, subset to only having columns of the
  correct value
}
\description{
This function is not standalone - it is called from the cr_extract function.
}
