% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_row.R
\name{subset_row}
\alias{subset_row}
\title{Subset the desired dataset, based on a row code.}
\usage{
subset_row(dataset, row)
}
\arguments{
\item{dataset}{The name of a cost report alpha or numeric dataset}

\item{row}{The number of the row}
}
\value{
A subset of the provided dataset, subset to only having rows of the
  correct value
}
\description{
This function is not standalone - it is called from the cr_extract function.
}
