% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_worksheet.R
\name{subset_worksheet}
\alias{subset_worksheet}
\title{Subset the desired dataset, based on a worksheet code.}
\usage{
subset_worksheet(dataset, worksheet)
}
\arguments{
\item{dataset}{The name of a cost report alpha or numeric dataset}

\item{worksheet}{The name of the workheet, converted to 7-character format}
}
\value{
A subset of the provided dataset, subset to only having worksheets of
  the correct value
}
\description{
This function is not standalone - it is called from the cr_extract function.
}
