% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicare-package.R
\docType{package}
\name{medicare-package}
\alias{medicare}
\alias{medicare-package}
\title{medicare: Tools for Obtaining and Cleaning Medicare Public Use Files}
\description{
Publicly available data from Medicare frequently requires extensive initial effort to extract desired variables and merge them; this package formalizes the techniques I've found work best. More information on the Medicare program, as well as guidance for the publicly available data this package targets, can be found on CMS's website covering publicly available data. See \url{https://www.cms.gov/data-research}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertgambrel/medicare}
  \item Report bugs at \url{https://github.com/robertgambrel/medicare/issues}
}

}
\author{
\strong{Maintainer}: Robert Gambrel \email{robert.gambrel@gmail.com}

Authors:
\itemize{
  \item Robert Gambrel \email{robert.gambrel@gmail.com}
}

}
\keyword{internal}
